Port PS2 SDL_main to header-only + SDL_RunApp()
